refactor(racingCar): domain, view 패키지 분리
