+++ src/main/kotlin/racingCar/controller/RacingGame.kt
@@ -1,4 +1,6 @@
-package racingCar
+package racingCar.controller
+
+import racingCar.model.Cars
 
 object RacingGame {
     fun playGame(cars: Cars) {
