+++ src/test/kotlin/racingCar/model/CarTest.kt
@@ -1,10 +1,11 @@
-package racingCar
+package racingCar.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingCar.model.Car
 
 class CarTest {
     private val car = Car("harry")
