+++ src/test/kotlin/racingCar/model/CarsTest.kt
@@ -1,8 +1,9 @@
-package racingCar
+package racingCar.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import racingCar.model.Cars
 
 class CarsTest {
     @DisplayName(value = "입력된 자동차 대수 만큼 자동차를 가진 리스트 생성")
